detect_rules/google-secops/requirements.txt
requests>=2.31.0
urllib3>=2.0.7
google-auth>=2.23.0
pyyaml>=6.0.1